requirements.txt
langchain
streamlit
OpenAI
huggingface_hub
python-dotenv